ISSD-1580 Move directory
