+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublishedAppsDashboard/Apps/AppCreationFlow/AppCreationFlowUtil.ts
