+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublishedAppsDashboard/Apps/AppCreationFlow/index.ts
@@ -3,7 +3,4 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export enum PaymentStatus {
-	PAID = 'completed',
-	PAYMENT_PENDING = 'pending',
-}
+export * from './AppCreationFlow';
